versionbits: Split out internal details into impl header
