+++ src/versionbits.h
@@ -24,18 +24,8 @@ static const int32_t VERSIONBITS_TOP_MASK = 0xE0000000UL;
 /** Total bits available for versionbits */
 static const int32_t VERSIONBITS_NUM_BITS = 29;
 
-/** BIP 9 defines a finite-state-machine to deploy a softfork in multiple stages.
- *  State transitions happen during retarget period if conditions are met
- *  In case of reorg, transitions can go backward. Without transition, state is
- *  inherited between periods. All blocks of a period share the same state.
- */
-enum class ThresholdState {
-    DEFINED,   // First state that each softfork starts out as. The genesis block is by definition in this state for each deployment.
-    STARTED,   // For blocks past the starttime.
-    LOCKED_IN, // For at least one retarget period after the first retarget period with STARTED blocks of which at least threshold have the associated bit set in nVersion, until min_activation_height is reached.
-    ACTIVE,    // For all blocks after the LOCKED_IN retarget period (final state)
-    FAILED,    // For all blocks once the first retarget period after the timeout time is hit, if LOCKED_IN wasn't already reached (final state)
-};
+/** Opaque type for BIP9 state. See versionbits_impl.h for details. */
+enum class ThresholdState : uint8_t;
 
 // A map that gives the state for blocks whose height is a multiple of Period().
 // The map is indexed by the block's parent, however, so all keys in the map
@@ -73,30 +63,6 @@ struct BIP9GBTStatus {
         bool gbt_force;
     };
     std::map<std::string, const Info, std::less<>> signalling, locked_in, active;
-};
-
-/**
- * Abstract class that implements BIP9-style threshold logic, and caches results.
- */
-class AbstractThresholdConditionChecker {
-protected:
-    virtual bool Condition(const CBlockIndex* pindex) const =0;
-    virtual int64_t BeginTime() const =0;
-    virtual int64_t EndTime() const =0;
-    virtual int MinActivationHeight() const { return 0; }
-    virtual int Period() const =0;
-    virtual int Threshold() const =0;
-
-public:
-    /** Returns the numerical statistics of an in-progress BIP9 softfork in the period including pindex
-     * If provided, signalling_blocks is set to true/false based on whether each block in the period signalled
-     */
-    BIP9Stats GetStateStatisticsFor(const CBlockIndex* pindex, std::vector<bool>* signalling_blocks = nullptr) const;
-    /** Returns the state for pindex A based on parent pindexPrev B. Applies any state transition if conditions are present.
-     *  Caches state from first block of period. */
-    ThresholdState GetStateFor(const CBlockIndex* pindexPrev, ThresholdConditionCache& cache) const;
-    /** Returns the height since when the ThresholdState has started for pindex A based on parent pindexPrev B, all blocks of a period share the same */
-    int GetStateSinceHeightFor(const CBlockIndex* pindexPrev, ThresholdConditionCache& cache) const;
 };
 
 /** BIP 9 allows multiple softforks to be deployed in parallel. We cache
